+++ src/main/java/com/github/brunothg/jappybird/object/Heli.java
@@ -1,4 +1,4 @@
-package com.github.brunothg.jappybird;
+package com.github.brunothg.jappybird.object;
 
 import java.awt.Color;
 import java.awt.Graphics2D;
